+++ tests/artifacts/devCloud01_10.7.2020.conf
