+++ tests/artifacts/mini_ucs.tar.gz
