+++ tests/devCloud_qkview.tests.ts
@@ -15,7 +15,7 @@ import assert from 'assert';
 import * as fs from 'fs';
 import * as path from 'path';
 
-import BigipConfig from '../ltm';
+import BigipConfig from '../src/ltm';
 import { logOutput } from './explosionOutput';
 
 /**
